# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Task {
  id: number;
  text: string;
  completed: boolean;
  note: string;
  position: number;
  /** Marks the task as scheduled for a future date — hidden from the active board, included in the daily report, and never carried over. */
  postedForFuture: boolean;
  /** ISO-8601 timestamp of when the task was created. */
  createdAt: string;
}

export interface TaskList {
  id: number;
  date: string;
  submitted: boolean;
  submittedAt: string | null;
  /** True if the user loaded the app today (signals a working day). Auto-submit at 23:59 only fires when this is true. */
  checkedIn: boolean;
  /** ISO-8601 timestamp of when the user first loaded the app today. */
  checkedInAt: string | null;
  tasks: Task[];
}

export interface CreateTaskListBody {
  /** Newline-separated task text */
  rawText: string;
}

export interface ToggleTaskBody {
  taskId: number;
}

export interface UpdateTaskNoteBody {
  taskId: number;
  note: string;
}

export interface UpdateTaskTextBody {
  taskId: number;
  text: string;
}

export interface SetPostedForFutureBody {
  taskId: number;
  postedForFuture: boolean;
}

export interface SubmitResult {
  success: boolean;
  message: string;
}

export interface TaskHistoryEntry {
  id: number;
  date: string;
  submitted: boolean;
  completedCount: number;
  totalCount: number;
}

export interface TaskHistoryList {
  entries: TaskHistoryEntry[];
}

export interface MatchedTaskSnippet {
  id: number;
  text: string;
  completed: boolean;
}

export interface TaskHistorySearchEntry {
  id: number;
  date: string;
  submitted: boolean;
  completedCount: number;
  totalCount: number;
  matchedTasks: MatchedTaskSnippet[];
}

export interface TaskHistorySearchResult {
  entries: TaskHistorySearchEntry[];
}

export interface ClearCompletedResult {
  success: boolean;
  deletedCount: number;
}

export interface UserSettings {
  /** Comma-separated email addresses */
  recipientEmails: string;
  /** When true, loading the app automatically marks you as checked in for the day. */
  autoCheckIn: boolean;
  /** When true, the server auto-submits at 23:59 for any day you checked in. */
  autoSubmit: boolean;
  /** Comma-separated JS day numbers (0=Sun…6=Sat) that count as work days. Default "1,2,3,4,5" = Mon–Fri. */
  workDays: string;
}

export interface UpdateSettingsBody {
  recipientEmails?: string;
  autoCheckIn?: boolean;
  autoSubmit?: boolean;
  workDays?: string;
}

export interface SetCheckedInBody {
  checkedIn: boolean;
}

export type SearchTaskHistoryParams = {
  /**
   * The search query (max 200 chars)
   */
  q: string;
  mode?: SearchTaskHistoryMode;
};

export type SearchTaskHistoryMode =
  (typeof SearchTaskHistoryMode)[keyof typeof SearchTaskHistoryMode];

export const SearchTaskHistoryMode = {
  name: "name",
  date: "date",
} as const;