# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get user settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          description: Unauthorized
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update user settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserSettings:
      type: object
      properties:
        recipientEmails:
          type: string
          description: Comma-separated email addresses
        autoCheckIn:
          type: boolean
          description: When true, loading the app automatically marks you as checked in for the day. Disable on holidays so auto-submit does not fire.
        autoSubmit:
          type: boolean
          description: When true, the server auto-submits at 23:59 for any day you checked in. Disable to always submit manually.
        workDays:
          type: string
          description: Comma-separated JS day numbers (0=Sun … 6=Sat) that count as work days. Default is "1,2,3,4,5" (Mon–Fri). Auto check-in and auto-submit are skipped on non-work days.
      required:
        - recipientEmails
        - autoCheckIn
        - autoSubmit
        - workDays
    UpdateSettingsBody:
      type: object
      properties:
        recipientEmails:
          type: string
        autoCheckIn:
          type: boolean
        autoSubmit:
          type: boolean
        workDays:
          type: string
...